# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Palletizer API",
    "description": "API headless de palettisation 3D et de chargement transport.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/orders/parse-csv": {
      "post": {
        "summary": "Parse Csv",
        "operationId": "parse_csv_api_v1_orders_parse_csv_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_parse_csv_api_v1_orders_parse_csv_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParseCsvResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/palletize": {
      "post": {
        "summary": "Palletize",
        "operationId": "palletize_api_v1_palletize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PalletizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PalletizeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/palletize/csv": {
      "post": {
        "summary": "Palletize Csv",
        "operationId": "palletize_csv_api_v1_palletize_csv_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_palletize_csv_api_v1_palletize_csv_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PalletizeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transport/load": {
      "post": {
        "summary": "Transport Load",
        "operationId": "transport_load_api_v1_transport_load_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransportLoadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransportLoadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_palletize_csv_api_v1_palletize_csv_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "orderId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Orderid"
          },
          "optimizationLevel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Optimizationlevel"
          },
          "minimumSupportRatio": {
            "type": "number",
            "title": "Minimumsupportratio",
            "default": 0.8
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_palletize_csv_api_v1_palletize_csv_post"
      },
      "Body_parse_csv_api_v1_orders_parse_csv_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_parse_csv_api_v1_orders_parse_csv_post"
      },
...
      "CsvImportErrorContract": {
        "properties": {
          "lineNumber": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linenumber"
          },
          "column": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Column"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "rawFragments": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rawfragments",
            "default": []
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "lineNumber",
          "code",
          "message"
        ],
        "title": "CsvImportErrorContract"
      },
      "CsvImportStatsContract": {
        "properties": {
          "ordersCount": {
            "type": "integer",
            "title": "Orderscount"
          },
          "palletFormats": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Palletformats"
          },
          "shippingModes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Shippingmodes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "ordersCount",
          "palletFormats",
          "shippingModes"
        ],
        "title": "CsvImportStatsContract"
      },
      "CsvImportWarningContract": {
        "properties": {
          "lineNumber": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linenumber"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "lineNumber",
          "message"
        ],
        "title": "CsvImportWarningContract"
      },
      "DimensionsContract": {
        "properties": {
          "length": {
            "type": "number",
            "title": "Length"
          },
          "width": {
            "type": "number",
            "title": "Width"
          },
          "height": {
            "type": "number",
            "title": "Height"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "length",
          "width",
          "height"
        ],
        "title": "DimensionsContract"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LegacyExpectedResultContract": {
        "properties": {
          "palletCount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Palletcount"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "palletCount"
        ],
        "title": "LegacyExpectedResultContract"
      },
...
      "OptimizationLevel": {
        "type": "string",
        "enum": [
          "fast",
          "thorough"
        ],
        "title": "OptimizationLevel"
      },
      "OptionsContract": {
        "properties": {
          "optimizationLevel": {
            "$ref": "#/components/schemas/OptimizationLevel",
            "default": "fast"
          },
          "minimumSupportRatio": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Minimumsupportratio",
            "default": 0.8
          },
          "globalRotationsEnabled": {
            "type": "boolean",
            "title": "Globalrotationsenabled",
            "default": true
          },
          "fragileMaxWeightOnTopKg": {
            "type": "number",
            "minimum": 0.0,
            "title": "Fragilemaxweightontopkg",
            "default": 0.0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "OptionsContract"
      },
      "OrderContract": {
        "properties": {
          "orderId": {
            "type": "string",
            "minLength": 1,
            "title": "Orderid"
          },
          "shippingMode": {
            "$ref": "#/components/schemas/ShippingMode"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/OrderLineContract"
            },
            "type": "array",
            "title": "Lines"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "orderId",
          "shippingMode",
          "lines"
        ],
        "title": "OrderContract"
      },
      "OrderLineContract": {
        "properties": {
          "lineNumber": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Linenumber"
          },
          "sku": {
            "type": "string",
            "minLength": 1,
            "title": "Sku"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "quantity": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Quantity"
          },
          "unit": {
            "type": "string",
            "title": "Unit",
            "default": "PIECE"
          },
          "dimensionsMm": {
            "$ref": "#/components/schemas/DimensionsMmContract"
          },
          "weightKg": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Weightkg"
          },
          "allowRotation": {
            "type": "boolean",
            "title": "Allowrotation",
            "default": true
          },
          "uprightOnly": {
            "type": "boolean",
            "title": "Uprightonly",
            "default": false
          },
          "fragile": {
            "type": "boolean",
            "title": "Fragile",
            "default": false
          },
          "stackable": {
            "type": "boolean",
            "title": "Stackable",
            "default": true
          },
          "maxSupportedWeightKg": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxsupportedweightkg"
          },
          "productGroup": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Productgroup"
          },
          "incompatibleGroups": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Incompatiblegroups",
            "default": []
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "lineNumber",
          "sku",
          "quantity",
          "dimensionsMm"
        ],
        "title": "OrderLineContract"
      },
...
      "PalletContract": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "default": ""
          },
          "lengthMm": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Lengthmm"
          },
          "widthMm": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Widthmm"
          },
          "maxHeightMm": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Maxheightmm"
          },
          "emptyPalletHeightMm": {
            "type": "number",
            "minimum": 0.0,
            "title": "Emptypalletheightmm",
            "default": 144.0
          },
          "maxHeightIncludesPallet": {
            "type": "boolean",
            "title": "Maxheightincludespallet",
            "default": true
          },
          "maxWeightKg": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxweightkg"
          },
          "overhangMm": {
            "type": "number",
            "minimum": 0.0,
            "title": "Overhangmm",
            "default": 0.0
          },
          "safetyGapMm": {
            "type": "number",
            "minimum": 0.0,
            "title": "Safetygapmm",
            "default": 0.0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "lengthMm",
          "widthMm",
          "maxHeightMm"
        ],
        "title": "PalletContract"
      },
      "PalletResultContract": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "config": {
            "$ref": "#/components/schemas/PalletConfigContract"
          },
          "totalWeightKg": {
            "type": "number",
            "title": "Totalweightkg"
          },
          "usableVolumeMm3": {
            "type": "number",
            "title": "Usablevolumemm3"
          },
          "volumeUsedMm3": {
            "type": "number",
            "title": "Volumeusedmm3"
          },
          "volumeUsageRatio": {
            "type": "number",
            "title": "Volumeusageratio"
          },
          "footprintUsageRatio": {
            "type": "number",
            "title": "Footprintusageratio"
          },
          "maxHeightUsedMm": {
            "type": "number",
            "title": "Maxheightusedmm"
          },
          "placedCartons": {
            "items": {
              "$ref": "#/components/schemas/PlacedCartonContract"
            },
            "type": "array",
            "title": "Placedcartons"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "index",
          "config",
          "totalWeightKg",
          "usableVolumeMm3",
          "volumeUsedMm3",
          "volumeUsageRatio",
          "footprintUsageRatio",
          "maxHeightUsedMm",
          "placedCartons"
        ],
        "title": "PalletResultContract"
      },
      "PalletToLoadContract": {
        "properties": {
          "palletResultIndex": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Palletresultindex"
          },
          "footprintLengthMm": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Footprintlengthmm"
          },
          "footprintWidthMm": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Footprintwidthmm"
          },
          "heightMm": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Heightmm"
          },
          "weightKg": {
            "type": "number",
            "minimum": 0.0,
            "title": "Weightkg"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "palletResultIndex",
          "footprintLengthMm",
          "footprintWidthMm",
          "heightMm",
          "weightKg"
        ],
        "title": "PalletToLoadContract"
      },
      "PalletizeRequest": {
        "properties": {
          "contractVersion": {
            "type": "string",
            "title": "Contractversion",
            "default": "1.0"
          },
          "order": {
            "$ref": "#/components/schemas/OrderContract"
          },
          "pallet": {
            "$ref": "#/components/schemas/PalletContract"
          },
          "options": {
            "$ref": "#/components/schemas/OptionsContract",
            "default": {
              "optimizationLevel": "fast",
              "minimumSupportRatio": 0.8,
              "globalRotationsEnabled": true,
              "fragileMaxWeightOnTopKg": 0.0
            }
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "order",
          "pallet"
        ],
        "title": "PalletizeRequest"
      },
      "PalletizeResponse": {
        "properties": {
          "contractVersion": {
            "type": "string",
            "title": "Contractversion",
            "default": "1.0"
          },
          "orderId": {
            "type": "string",
            "title": "Orderid"
          },
          "engineVersion": {
            "type": "string",
            "title": "Engineversion"
          },
          "durationMs": {
            "type": "number",
            "title": "Durationms"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "totalCartonsCount": {
            "type": "integer",
            "title": "Totalcartonscount"
          },
          "placedCartonsCount": {
            "type": "integer",
            "title": "Placedcartonscount"
          },
          "unplacedCartonsCount": {
            "type": "integer",
            "title": "Unplacedcartonscount"
          },
          "palletsCount": {
            "type": "integer",
            "title": "Palletscount"
          },
          "globalVolumeUsageRatio": {
            "type": "number",
            "title": "Globalvolumeusageratio"
          },
          "totalWeightKg": {
            "type": "number",
            "title": "Totalweightkg"
          },
          "pallets": {
            "items": {
              "$ref": "#/components/schemas/PalletResultContract"
            },
            "type": "array",
            "title": "Pallets"
          },
          "unplacedCartons": {
            "items": {
              "$ref": "#/components/schemas/UnplacedCartonContract"
            },
            "type": "array",
            "title": "Unplacedcartons"
          },
          "legacyExpectedResult": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LegacyExpectedResultContract"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "orderId",
          "engineVersion",
          "durationMs",
          "warnings",
          "totalCartonsCount",
          "placedCartonsCount",
          "unplacedCartonsCount",
          "palletsCount",
          "globalVolumeUsageRatio",
          "totalWeightKg",
          "pallets",
          "unplacedCartons"
        ],
        "title": "PalletizeResponse"
      },
      "ParseCsvResponse": {
        "properties": {
          "contractVersion": {
            "type": "string",
            "title": "Contractversion",
            "default": "1.0"
          },
          "orders": {
            "items": {
              "$ref": "#/components/schemas/ParsedCsvOrderContract"
            },
            "type": "array",
            "title": "Orders"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/CsvImportErrorContract"
            },
            "type": "array",
            "title": "Errors"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/CsvImportWarningContract"
            },
            "type": "array",
            "title": "Warnings"
          },
          "totalRows": {
            "type": "integer",
            "title": "Totalrows"
          },
          "acceptedRows": {
            "type": "integer",
            "title": "Acceptedrows"
          },
          "rejectedRows": {
            "type": "integer",
            "title": "Rejectedrows"
          },
          "stats": {
            "$ref": "#/components/schemas/CsvImportStatsContract"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "orders",
          "errors",
          "warnings",
          "totalRows",
          "acceptedRows",
          "rejectedRows",
          "stats"
        ],
        "title": "ParseCsvResponse"
      },
      "ParsedCsvOrderContract": {
        "properties": {
          "orderId": {
            "type": "string",
            "title": "Orderid"
          },
          "shippingMode": {
            "$ref": "#/components/schemas/ShippingMode"
          },
          "pallet": {
            "$ref": "#/components/schemas/PalletContract"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/OrderLineContract"
            },
            "type": "array",
            "title": "Lines"
          },
          "legacyPalletCount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legacypalletcount"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "orderId",
          "shippingMode",
          "pallet",
          "lines"
        ],
        "title": "ParsedCsvOrderContract"
      },
...
      "RejectionCode": {
        "type": "string",
        "enum": [
          "DIMENSIONS_EXCEED_PALLET",
          "HEIGHT_EXCEEDED",
          "WEIGHT_EXCEEDED",
          "ROTATION_FORBIDDEN",
          "STACKING_CONSTRAINT",
          "NO_STABLE_POSITION",
          "INVALID_DATA",
          "INCOMPATIBLE_GROUP"
        ],
        "title": "RejectionCode"
      },
      "ShippingMode": {
        "type": "string",
        "enum": [
          "sea",
          "air",
          "road",
          "unknown"
        ],
        "title": "ShippingMode"
      },
      "TransportLoadRequest": {
        "properties": {
          "contractVersion": {
            "type": "string",
            "title": "Contractversion",
            "default": "1.0"
          },
          "pallets": {
            "items": {
              "$ref": "#/components/schemas/PalletToLoadContract"
            },
            "type": "array",
            "title": "Pallets"
          },
          "vehicle": {
            "$ref": "#/components/schemas/VehicleContract"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "pallets",
          "vehicle"
        ],
        "title": "TransportLoadRequest"
      },
      "TransportLoadResponse": {
        "properties": {
          "contractVersion": {
            "type": "string",
            "title": "Contractversion",
            "default": "1.0"
          },
          "vehicles": {
            "items": {
              "$ref": "#/components/schemas/VehicleLoadResultContract"
            },
            "type": "array",
            "title": "Vehicles"
          },
          "unassignedPalletIndexes": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Unassignedpalletindexes"
          },
          "vehiclesNeeded": {
            "type": "integer",
            "title": "Vehiclesneeded"
          },
          "palletsLoadable": {
            "type": "integer",
            "title": "Palletsloadable"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "vehicles",
          "unassignedPalletIndexes",
          "vehiclesNeeded",
          "palletsLoadable"
        ],
        "title": "TransportLoadResponse"
      },
      "UnplacedCartonContract": {
        "properties": {
          "instanceId": {
            "type": "string",
            "title": "Instanceid"
          },
          "sku": {
            "type": "string",
            "title": "Sku"
          },
          "dimensionsMm": {
            "$ref": "#/components/schemas/DimensionsContract"
          },
          "code": {
            "$ref": "#/components/schemas/RejectionCode"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "weightKg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weightkg"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "instanceId",
          "sku",
          "dimensionsMm",
          "code",
          "message"
        ],
        "title": "UnplacedCartonContract"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VehicleContract": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "default": "vehicule"
          },
          "innerLengthMm": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Innerlengthmm"
          },
          "innerWidthMm": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Innerwidthmm"
          },
          "innerHeightMm": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Innerheightmm"
          },
          "maxPayloadKg": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Maxpayloadkg"
          },
          "allowPalletRotationFloor": {
            "type": "boolean",
            "title": "Allowpalletrotationfloor",
            "default": true
          },
          "allowPalletStacking": {
            "type": "boolean",
            "title": "Allowpalletstacking",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "innerLengthMm",
          "innerWidthMm",
          "innerHeightMm",
          "maxPayloadKg"
        ],
        "title": "VehicleContract"
      },
      "VehicleLoadResultContract": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "placements": {
            "items": {
              "$ref": "#/components/schemas/LoadedPalletPlacementContract"
            },
            "type": "array",
            "title": "Placements"
          },
          "usedFloorAreaRatio": {
            "type": "number",
            "title": "Usedfloorarearatio"
          },
          "usedWeightKg": {
            "type": "number",
            "title": "Usedweightkg"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "index",
          "placements",
          "usedFloorAreaRatio",
          "usedWeightKg"
        ],
        "title": "VehicleLoadResultContract"
      }
    }
  }
}